Removed validateEffectiveDate from shared because individual properties and re-connected to local validation.js
